### 2018-03-04 #201 @cimentadaj
submission: ess
It allows users to download and explore data from the European Social Survey directly into R. Broadly speaking, it has two family of functions. The `ess_*` family, which downloads data from specific waves and countries and the `show_*` functions which allows the users to check which waves/countries are available.

```
Package: ess
Title: Download Data from the European Social Survey on the Fly
Version: 0.1.1
Authors@R: person("Jorge", "Cimentada", email = "cimentadaj@gmail.com", role = c("aut", "cre"))
Description: Download data from the European Social Survey directly from their website <http://www.europeansocialsurvey.org/>. There are two family of functions that allow you to download and interactively check all countries and rounds available.
    xml2,
    httr,
    haven,
    rvest,
    tibble
    covr
Note: The ess package is named to match the European Social Survey (ESS) hosted at http://www.europeansocialsurvey.org/. It is unrelated to the Emacs Speaks Statistics (ESS) project, an emacs-based Integrated Development Environment hosted at https://ess.r-project.org
```

https://github.com/cimentadaj/ess

[e.g., "data extraction, because the package parses a scientific data file format"]

data retrieval, because the main objective of the package is to download a set of datasets widely used in the social sciences.

Mainly academics working in disciplines such as Economics, Demography, Psychology, Political Science and Sociologists. Hundreds of papers have been written using the European Social Survey, a survey carried over in over 30 countries for over 15 years that asks citizens everything from their health status to their political preferences. This package will allow full reproducibility for many of the academics using the survey and equally important, will remove the burden of manually exploring/downloading the datasets from the ESS website.

The [lodown](https://github.com/ajdamico/lodown) package (not on CRAN) has one function for downloading the data as 'Stata' or 'SPSS' files for only specific waves. The package is not intended specifically for the ESS data. This package differs because it allows the user to download the files directly into R selecting different wave/country combinations. It also supports downloading the files in 'Stata', 'SPSS' and in 'SAS' and offers a range of functions for exploring the available countries/waves directly from R.

It succeeds with no errors/warnings.

It conforms.

I don't know anyone who has experience with this type of packages but I assume anyone with relevant experience in web scraping is adequate. The workhorse of the package is scraping the ESS website, reading it with the `haven` package and putting the data in a suitable format for R.
