+++ NEWS.md
@@ -1,5 +1,6 @@
 ## ess 0.1.1.9000
 
+No changes so far
 
 ## ess 0.1.1 (2018-03-05)
 
