### 2018-03-08 #201 @maelle
### Editor checks:

---

#### Editor comments

Thanks for your submission @cimentadaj! :smile_cat: The tests are now finished. 😌  I'm currently looking for a second reviewer (thanks @leeper for accepting to review!).

Please add this badge to `ess` README

Here are a few comments/questions.

* Just curious, why is there a note about the package name in DESCRIPTION, was that demanded by CRAN?

* The link from the pkgdown website is missing from the repo description.

* `goodpractice::gp` indicates a few long code lines. It seems that they are error messages so it is fine. You can escape them with # nolint cf https://github.com/jimhester/lintr#project-configuration

* Please document the fact that one needs to set "your_email" as an environment variable. I moreover think its name is a bit confusing, it is *my* email :wink: maybe it should be "ess_account_email"?

* I'd recommend you to look for the email by default like what is done [in this package](https://github.com/ropensci/monkeylearn/blob/master/R/utils.R#L152)

---

Reviewers: @leeper
Due date: 2018-03-29
