### 2018-03-09 cimentadaj/ess#20 @leeper
Add package-level documentation
Currently, the package does not have a package-level documentation page:

```R
> library("ess")
> ? ess
No documentation for ‘ess’ in specified packages and libraries:
you could try ‘??ess’
```

Having this can be useful for giving new users an obvious landing page.
