### 2018-03-09 cimentadaj/ess#21 @leeper
Clarify whether show_rounds_country() is an AND or OR logical
The `show_rounds_country()` function takes a numeric vector of ESS rounds (as integers) and returns participating countries. It's not obvious from [the documentation](https://github.com/cimentadaj/ess/blob/master/R/show_any_rounds.R#L3-L4), however, whether the countries that are returned are those that participated in *any* of those rounds or the set of countries that participated in *all* of those rounds. Some clarity might be useful or, alternative, two functions (or a new argument to this function) to disambiguate might be appropriate.
