### 2018-03-09 cimentadaj/essurvey#22 @leeper
Distinguish download and import functionality as distinct functions
The `ess_rounds()` function, and related functions, can both import and simply down a data file. As far as I can tell when importing, the data file is not saved to the local directory. It seems you may want to distinguish these by creating separate `download_rounds()` and `import_rounds()` functions since their behavior is quite different - the former is only used for a side effect.

This then has an impact on a feature of `ess_rounds()` that seems unnecessary for most users: namely, the ability to download different file formats. If the users is only importing into R, it shouldn't matter what file format they download so it might be safe to remove this option entirely from an `import_rounds()` function.
