### 2018-03-09 cimentadaj/ess#23 @leeper
Make 'your_email' a global variable
It would be great to do one of the following:

- [ ] Have `ESS_EMAIL` as an environment variable so it can be set globally and not within each call
- [ ] If not that, then have a function like `set_email()` that sets a hidden global variable for a given session
