### 2018-03-09 cimentadaj/ess#24 @leeper
Use HTTPS
It looks like the package currently using insecure HTTP calls to authenticate the user, for example:

- https://github.com/cimentadaj/ess/blob/master/R/download_format.R#L67
- https://github.com/cimentadaj/ess/blob/master/R/download_format.R#L71

As well as in several other places. It appears the ESS website supports HTTPS, so it would be good to switch to that if possible. While an email address isn't exactly a super sensitive piece of data, users should be protected to the extent possible. If that's not possible, I would flag it explicitly in the README that this is passing the email address in the open.
