### 2018-03-09 cimentadaj/ess#25 @leeper
Use match.arg() where appropriate
There are a few places where constructions like [the following](https://github.com/cimentadaj/ess/blob/master/R/download_format.R#L12-L14) are used:

```R
# Check if the format is either 'stata', 'spss' or 'sas'.
  if(!format %in% c('stata', 'spss', 'sas')) {
    stop("Format not available. Only 'stata', 'spss', or 'sas'")
  }
```

This is a great opportunity to use the `match.arg()` function, which will produce basically the same error message.
