### 2018-03-09 cimentadaj/ess#26 @leeper
Mention survey weights
One of the big questions when using survey data in R is how best to handle weighting. The {survey} package provides very good support for this and most users would need to know (if they don't already) that they need to dump the ESS data into a `survey::svydesign()` object before doing any analysis. It might be useful to have a second vignette showing how to do this and perform some analyses, or at least mention survey weighting in the README, vignette, and package-level documentation.
