### 2018-03-09 #201 @leeper
This package is a very easy-to-use tool for download ESS data. It will be of great benefit to ESS users in both a teaching and research context. In the latter case, it will no doubt help considerably with reproducibility of analyses using the ESS. The code is generally well-written and the documentation is mostly excellent. Both the README and vignette provide useful introductions to the package and explain all functionality.

I have flagged seven issues I would like to see responses to:

- [ ] https://github.com/cimentadaj/ess/issues/20
- [ ] https://github.com/cimentadaj/ess/issues/21
- [ ] https://github.com/cimentadaj/ess/issues/22
- [ ] https://github.com/cimentadaj/ess/issues/23
- [ ] https://github.com/cimentadaj/ess/issues/24
- [ ] https://github.com/cimentadaj/ess/issues/25
- [ ] https://github.com/cimentadaj/ess/issues/26

These are all suggestions for improvement rather than obvious bugs or problems with the package. The only other suggestion I would have is that this package may significantly increase interest in the ESS among R users and therefore it may be valuable to expand the examples in the README or add an additional vignette that showcases some of the really interesting substantive data available in the datasets. It would be great to see a few tables or graphs showing some interesting patterns. Not necessary, of course, but that might help sell R users on the ESS in the way that the package is already helping sell ESS users on R.

Great package.
