### 2018-03-09 #201 @maelle
Wow thanks a lot for your extremely rapid and good review @leeper 😮 👏

Could you please fill and add [the review template](https://github.com/ropensci/onboarding/blob/master/issue_template.md) to your comment? Thanks in advance.
