### 2018-03-09 ropensci/essurvey#20 @maelle
Note that it can be added very easily by calling `usethis::use_package_doc()`
