### 2018-03-09 #201 @maelle
Just a suggestion as a reaction to @leeper's excellent suggestion to add examples, I tend to think minimal READMEs make more sense with a pkgdown website (especially if the README has links to articles/vignettes), so adding a vignette or several ones might be better than expanding the README.
