### 2018-03-09 #201 @maelle
Thanks for agreeing to review this package @nujcharee! :smile_cat: Your reviews is due on 2018-03-30.

As a reminder here are links to the recently updated [reviewing](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md) and [packaging](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md) guides and to the [review template](https://raw.githubusercontent.com/ropensci/onboarding/master/reviewer_template.md).

@cimentadaj you can wait for @nujcharee's review before responding if you wish.
