### 2018-03-11 cimentadaj/ess#27 @nujcharee
README file - missing library
Users may be beneficial from the below code to include "library(tidyverse)" to avoid encountering an error "could not find function %>%"

one_round <-
  ess_rounds(1, "your_email@email.com") %>%
  recode_missings()
