### 2018-03-11 #201 @nujcharee
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

The ESS package provides programmatic access to European Social Survey published on http://www.europeansocialsurvey.org/downloadwizard/. Users can directly download the surveys from the website in 3 different formats: SPSS, STATA and CSV. The ESS package is compact and easy to use with 2 main family functions: show_* and ess_*. It also offers functions that handle missing values.

The package doesn't require dependencies. It uses a valid registered email for authentication. The vignettes gives an introduction of what ESS is and contains a link to the website. It also demonstrates common usage and are easy to follow.

I installed and ran it successfully on Windows 7 (R-3.4.2) and Mac OS X (R-3.3.2)
I used test_package("ess") and received the following message   OK: 38 SKIPPED: 23 FAILED: 0

#### Possibilities for improvement
* README file - maybe a good idea to include library(tidyverse) in the code snippet below to avoid error prior "could not find function "%>%"
[ ] https://github.com/cimentadaj/ess/issues/27

library(tidyverse)
one_round <-
  ess_rounds(1, "your_email@email.com") %>%
  recode_missings()
```
* In my opinion, for those you aren't familiar with Social Survey, they might be really beneficial from a brief description of what the term "round" mean and how this is used in the context of social survey.
* Opportunities to apply global variables for "email address"
* Opportunities to merge functions "ess_all_rounds()" and "ess_rounds()" where "all" is default if country variables aren't passed.
* Making the country_name argument case insensitive is a potential improvement.

### Final note
Great package, I can see the potential for it to become very popular among social science professionals.
