### 2018-03-12 #201 @jimhester
I have no background in social sciences so have no idea how ubiquitous the acronym ESS for European Social Survey is, but in the R / S community there is long precedent for ESS standing for [Emacs Speaks Statistics](https://ess.r-project.org/), dating back to at least [1994](ftp://stat.ethz.ch/Mail-archives/ess-help/1994-1996/), before R existed. Emacs Speaks Statistics is also quite popular for emacs R users, so this is not a niche issue.

I am actually astonished that this package was allowed on CRAN with this name and would really encourage considering changing the name to something more descriptive; maybe `essurvey` would suffice, you could retain the `ess_` shorthand for the internal functions.
