### 2018-03-14 adb6313
Add rOpensci badge
