### 2018-03-14 #201 @cimentadaj
Everyone, thanks all the suggestions. I will go one by one. @maelle

* Just curious, why is there a note about the package name in DESCRIPTION, was that demanded by CRAN?

This is because it has conflicts with Emacs Speaks Statistics and the Emacs authors suggested that I add it as a note and they will add it as well to their R package. In fact, the point raised by @jimhester has been raised by several other R users in this email [thread](http://r.789695.n4.nabble.com/R-pkgs-Release-of-ess-0-0-1-td4746540.html). Among social scientists the ESS acronym is widely known and would be the first thing that comes to mind, yet not many of these scientists are R users, whereas Emacs Speaks Statistics is widely used/known in the R community. I've decided that it is best for the package to be renamed and rereleased in the future. In fact, I was thinking of using the same name @jimhester suggested, `essurvey`. I don't know how this affects the rOpensci submission, let me know if something needs to be changed. I will make the name change effective when I have time, somewhere in the next few weeks.

* The link from the pkgdown website is missing from the repo description.

Done 👍

* goodpractice::gp indicates a few long code lines. It seems that they are error messages so it is fine. You can escape them with # nolint cf https://github.com/jimhester/lintr#project-configuration

Done 👍

Please document the fact that one needs to set "your_email" as an environment variable. I moreover think its name is a bit confusing, it is my email 😉 maybe it should be "ess_account_email"? I'd recommend you to look for the email by default like what is done in this package

Currently `your_email` is not an environment variable but purely an argument. Following the comments of @leeper I think a better approach would be set it as environment/option variables but also allow the email as an argument (for backward dependency), encouraging the user not to use the email as an argument; this would be explained in detail in the vignette, which might cause confusion for ESS users not familiar with R. The approach would be similar but slightly different from the package you suggested.
