### 2018-03-14 #201 @cimentadaj
@leeper I will review your suggestions in the coming days. I think that making a more 'practical' vignette is also a good idea. I will work on it.
