### 2018-03-14 #201 @cimentadaj
@nujcharee thank you very much for your comments! I agree with your suggestions and will implement them soon.  In fact, I think you also raised some common points that overlap with what @leeper suggested. Thanks!
