### 2018-03-14 ropensci/essurvey#25 @cimentadaj
I hadn't thought about it! Thanks. I'm applying it to the case of matching the `format` argument through out the package. However, in the other cases  prefer to have a separate `stop` message, like [here](https://github.com/cimentadaj/ess/blob/master/R/ess_country_url.R#L10-L23). This is because it checks whether the supplied country name is available in the ESS website and instead of `match.arg` returning all the country names (long) I `stop` with a clear message suggesting the user to check `show_countries()` for the correct country name. Applies to rounds (and themes), with an even more explicit error.

There are other instances where it make sense such as [here](https://github.com/cimentadaj/ess/blob/master/R/recode_missings.R#L132-L138) but it doesn't work as expect because it needs to match every code and throw an error if it doesn't. For example `match.arg(c("123", "Refusal"), c("Don't Know", "Refusal"), TRUE)` only returns `Refusal` but doesn't raise an error. So in these other cases I also think the error message is more strict.  See https://github.com/cimentadaj/ess/commit/0336d37ae09003f25c60ea769a9fc685d67f03e7.
