### 2018-03-14 #201 @cimentadaj
Oops, answered first on the issue as per . Copying here.

As per https://github.com/cimentadaj/ess/issues/25

I hadn't thought about it! Thanks. I'm applying it to the case of matching the `format` argument through out the package. However, in the other cases  prefer to have a separate `stop` message, like [here](https://github.com/cimentadaj/ess/blob/master/R/ess_country_url.R#L10-L23). This is because it checks whether the supplied country name is available in the ESS website and instead of `match.arg` returning all the country names (long) I `stop` with a clear message suggesting the user to check `show_countries()` for the correct country name. Applies to rounds (and themes), with an even more explicit error.
