### 2018-03-14 631f831
Exclude long error lines from lintr
