+++ tests/testthat/test-country.R
@@ -110,5 +110,5 @@ test_that("Download country files with other non-stata format", {
 
 test_that("Specify 'sas' for reading ess data throws error",{
   expect_error(ess_country("Denmark", 1, your_email, format = "sas"),
-               "You cannot read SAS but only 'spss' and 'stata' files with this function")
+               "You cannot read SAS but only 'spss' and 'stata' files with this function") #nolint
 })
