+++ tests/testthat/test-rounds.R
@@ -105,5 +105,5 @@ test_that("Download round files with other non-stata format", {
 
 test_that("Specify 'sas' for reading ess data throws error",{
   expect_error(ess_rounds(1, your_email, format = "sas"),
-               "You cannot read SAS but only 'spss' and 'stata' files with this function")
+               "You cannot read SAS but only 'spss' and 'stata' files with this function") # nolint
 })
