### 2018-03-14 #201 @jimhester
In terms of the particulars of renaming the package, I think the best way to do it is to submit the package to CRAN with the new name, and in your comments to CRAN in the submission say you want to rename the package and archive the old name.
