### 2018-03-14 #201 @maelle
@cimentadaj changing names doesn't impact the submission. I know that because my `Ropenaq` package was renamed `ropenaq` during its onboarding 😉 (I didn't have it on CRAN but you have @jimhester's guidance above, by the way thanks Jim for chiming in!)
