### 2018-03-15 57656ff
Switch auth to https over http
