### 2018-03-15 #201 @cimentadaj
As per @leeper in https://github.com/cimentadaj/ess/issues/24

Added in https://github.com/cimentadaj/ess/commit/57656ffdcbac6e156752602009210238e7e9ea4f

Good suggestion, I had no idea of the difference between HTTPS and HTTP.
