### 2018-03-18 #201 @maelle
@cimentadaj great job! 👌  Just a note about the process, it'd be easier for future readers of the thread to group your answers and to write here what the issue was about so that one doesn't need to click. Moreover if you want to link a commit to an issue in the commit message you can write something like "blabla closes #" with the issue number (or just "see" instead of "close" if you don't want to close it yet)
