### 2018-03-19 76007c5
Make clear that show_rounds_country returns ALL rather than OR
