### 2018-03-19 ropensci/essurvey#21 @cimentadaj
See https://github.com/cimentadaj/ess/commit/76007c5ac35161629a3eefed4a941a9a726dee0d where I improved documentation as to make it clear with some examples. Adding a `OR` type of function could be handy, and something that I will add for next releases yet I don't think it's urgent. Users can find a work around around along theses lines `purrr::map(ess::show_rounds(), ess::show_rounds_country)`.
