### 2018-03-20 ed0df5c
Add package level documentation. Close #20
