+++ R/ess-package.R
@@ -1,0 +1,4 @@
+#' @keywords internal
+#' @details For detailed examples visit the package
+#'  website at \link{https://cimentadaj.github.io/ess/}
+"_PACKAGE"
