### 2018-03-20 40dbf67
Add package level documentation. Close #20
