+++ R/ess-package.R
@@ -1,0 +1,4 @@
+#' @keywords internal
+#' @details For detailed examples visit the package
+#'  website at \url{https://cimentadaj.github.io/ess/}
+"_PACKAGE"
