### 2018-03-20 d83f728
Add notes on analyzing ess data. Close #26
