+++ R/ess-package.R
@@ -1,4 +1,9 @@
 #' @keywords internal
-#' @details For detailed examples visit the package
-#'  website at \url{https://cimentadaj.github.io/ess/}
+#' @details Note that this package is for downloading data from the ESS survey.
+#' For analyzing the data, the user should consider the weights and sampling
+#' design of each country/round combination as well as sampling between round.
+#' For some examples, check out the work of Anthony Damico and Daniel Oberski 
+#' \href{http://asdfree.com/european-social-survey-ess.html}{here}. For detailed
+#' examples on how to explore/download data using package, visit the package website
+#' at \url{https://cimentadaj.github.io/ess/}
 "_PACKAGE"
