### 2018-03-20 be8722e
Add changes to NEWS
