### 2018-03-21 11fcf11
Add set_email + change arg name of your_email
