### 2018-03-24 f1c7300
Add meaning on term round in readme/vignette
