+++ vignettes/intro_ess.Rmd
@@ -19,7 +19,7 @@ knitr::opts_chunk$set(
 
 ## Introduction
 
-Using the `ess` package is fairly easy. There are are two main families of functions: `ess_*` and `show_*`. They each complement each other and allow the user to almost never have to go to the European Social Survey (ESS) website. The only scenario where you need to enter the ESS website is to validate your email. If you haven't registered, create an account at http://www.europeansocialsurvey.org/user/new
+Using the `ess` package is fairly easy. There are are two main families of functions: `ess_*` and `show_*`. They each complement each other and allow the user to almost never have to go to the European Social Survey (ESS) website. The only scenario where you need to enter the ESS website is to validate your email. If you haven't registered, create an account at http://www.europeansocialsurvey.org/user/new. For those unfamiliar with the ESS, this vignette uses the term rounds, here a synonym of waves to denote the same survey in different time points.
 
 Once you register visit your email account to validate the account and you're ready to access the data.
 
