### 2018-03-24 1275038
Change name of environ var for tests
