+++ tests/testthat/test-country.R
@@ -1,13 +1,13 @@
 # Environment variables from Travis CI
 
-your_email <- Sys.getenv("your_email")
+ess_email <- Sys.getenv("ess_email")
 
 test_that("ess_country for one round", {
   
   skip_on_cran()
   
   # Test for only one wave
-  wave_one <- ess_country("Denmark", 1, your_email, format = 'stata')
+  wave_one <- ess_country("Denmark", 1, ess_email, format = 'stata')
   
   # check is list
   expect_is(wave_one, "data.frame")
@@ -24,7 +24,7 @@ test_that("ess_country for all rounds of a country", {
   skip_on_cran()
   
   # Test for all rounds
-  all_rounds <- ess_country("Netherlands", 1:7, your_email)
+  all_rounds <- ess_country("Netherlands", 1:7, ess_email)
   
   # check is list
   expect_is(all_rounds, "list")
@@ -54,7 +54,7 @@ test_that("Test that downloading files is working fine", {
   expect_message(downloads <-
                    ess_country("Austria",
                                1:which_rounds,
-                               your_email,
+                               ess_email,
                                only_download = TRUE,
                                output_dir = tempdir()
                    ),
@@ -86,7 +86,7 @@ test_that("Test if only_download is TRUE, output_dir should be valid", {
   # Here output_dir is set to NULL
   expect_error(ess_country("Austria",
                            1:which_rounds,
-                           your_email,
+                           ess_email,
                            only_download = TRUE,
                            output_dir = NULL))
 })
@@ -95,7 +95,7 @@ test_that("Download country files with other non-stata format", {
   skip_on_cran()
   
   # Test for only one wave
-  wave_one <- ess_country("Denmark", 1, your_email, format = "spss")
+  wave_one <- ess_country("Denmark", 1, ess_email, format = "spss")
   
   # check is list
   expect_is(wave_one, "data.frame")
@@ -109,6 +109,6 @@ test_that("Download country files with other non-stata format", {
 })
 
 test_that("Specify 'sas' for reading ess data throws error",{
-  expect_error(ess_country("Denmark", 1, your_email, format = "sas"),
+  expect_error(ess_country("Denmark", 1, ess_email, format = "sas"),
                "You cannot read SAS but only 'spss' and 'stata' files with this function") #nolint
 })
