### 2018-03-24 #201 @cimentadaj
Here I will respond to @nujcharee's suggestions.

In https://github.com/cimentadaj/ess/issues/27, @nujcharee suggested adding `library(tidyverse)` to the `README`. Implemented in https://github.com/cimentadaj/ess/commit/25628bc99ff679f27ce428af105d425fd60e277e.

Some other suggestions.

In my opinion, for those which aren't familiar with Social Survey, it might be really beneficial from a brief description of what the term "round" mean and how this is used in the context of social survey. Thank you for this! Implemented it in https://github.com/cimentadaj/ess/commit/f1c730069b3f9ecd63ee56c0f735af9c7f7f1441

**Opportunities to apply global variables for "email address"**: applied in https://github.com/cimentadaj/ess/commit/11fcf116d1f821dcb2aeef2ed154e67cf2c32aa0 from https://github.com/cimentadaj/ess/issues/23

**Opportunities to merge functions "ess_all_rounds()" and "ess_rounds()" where "all" is default if country variables aren't passed.** I'm not sure I get this @nujcharee .

Making the country_name argument case insensitive is a potential improvement.
