### 2018-03-24 ropensci/essurvey#22 @cimentadaj
This change makes a lot of sense and I think is the right way to approach this. In fact, the This change makes a lot of sense and I think is the right way to approach this. In fact, the implementation should be straight forward. However, I'm concerned about backward compatibility and the fact that I plan to change the name of the package to `essurvey` in the near future.

My main concern is having **both** the breaking compatibility change in the `ess` package and then changing the name of the package in a short period of time. I would want to give some time to users to communicate the API changes and let them know that the functions will be deprecated without also swamping them with a new package name (let's remember that the users of this package are not necessarily R users but people who come to R for specific things). I've thought a lot about this and I think this could  be one solution:

* Release the package with all the changes from the rOpensci (including this change) review as the new package name `essurvey`.

* This new release will contain the `import_*` and `download_*` functions respectively and provide a deprecated message for all `ess_*` functions. This specifies that the functions will be removed from the package altogether in the next release (3 months, more or less)

* Publicising the package

* Remove `ess_*` functions 3-4 months later.

I'm really hesitant on this step so I'd like to hear your comments. Perhaps it's not the best idea to apply these changes in the first place because the breaking changes are not extremely necessary.

I'd really like to hear your feedback on this particular point.
