### 2018-03-24 ropensci/essurvey#23 @cimentadaj
I think this is very useful feedback! I implemented this in https://github.com/cimentadaj/ess/commit/11fcf116d1f821dcb2aeef2ed154e67cf2c32aa0

A summary of the changes:
* Add `set_email("email@email.com") function
* Change the examples/vignette/READMe to a workflow using `set_email` rather than showing the email.
* Change the `your_email` argument name to `ess_email` in all `ess_*` functions.

All [tests pass](https://travis-ci.org/cimentadaj/ess/builds/356540744?utm_source=github_status&utm_medium=notification), so it should be good to do. (Again, very sorry that the commits seem to rewrite whole documents. Something weird is happening when making changes using the Ubuntu terminal in Windows 10.)
