### 2018-03-24 ropensci/essurvey#24 @cimentadaj
See https://github.com/cimentadaj/ess/commit/57656ffdcbac6e156752602009210238e7e9ea4f for the implementation.

Good suggestion, I had no idea of the difference between HTTPS and HTTP. Let me know if this is fine, to close the issue.
