### 2018-03-24 ropensci/essurvey#22 @leeper
My advice would be to deprecate the current function using [`.Deprecated()`](https://stat.ethz.ch/R-manual/R-devel/library/base/html/Deprecated.html) and then remove it sometime in the future. As long as you're changing the package name, that seems like a fine time to deprecate (or even remove as users of the new package will likely be expecting API changes).
