### 2018-03-25 #201 @nujcharee
I think in terms of the function “all” I think personally that “all” function can be obselete if it can be embedded in ess_round function when countries aren’t specified.

I made a suggestion as a recommendation. Looking forward to seeing this on CRAN soon good luck.
