### 2018-03-25 #201 @cimentadaj
Thanks a lot @nujcharee! @maelle, if the package gets accepted we need to sort out the issue with name of the package. Right now the name of the package is `ess` both in CRAN and on Github. During this week I'm going to finish all of the changes from the review and change the Github package name to `essurvey` and resubmit to CRAN in the next weeks. Perhaps the package can be `essurvey` in rOpensci while I resubmit it to CRAN with that name.
