+++ man/import_rounds.Rd
@@ -22,10 +22,11 @@ If you haven't registered in the ESS website, create an account at
 \url{http://www.europeansocialsurvey.org/user/new}. A prefered method is to login
 through \code{\link{set_email}}.}
 
-\item{format}{the format from which to download the data. Can either be 'stata', 'spss' or 'sas',
-with 'stata' as default. When \code{only_download} is set to TRUE, the data will be downloaded in
-the \code{format} specified. If \code{only_download} is FALSE, the data is downloaded and read
-from the specified \code{format} (only 'spss' and 'stata' supported, see details).}
+\item{format}{the format from which to download the data. Can either be 'stata',
+'spss' or 'sas', with 'stata' as default. When using \code{import_rounds} the
+data will be downloaded and read in the \code{format} specified (only 'spss'
+and 'stata' supported, see details). For \code{download_rounds}, the data is
+downloaded from the specified \code{format}.}
 
 \item{output_dir}{a character vector with the output directory in case you want to only download the files using
 the \code{download_rounds}. Defaults to your working directory. This will be interpreted as
@@ -53,7 +54,7 @@ different handling of the encoding of some questions. This option was preserved
 so that the user
 can switch between formats if any encoding errors are found in the data. For more
 details see the dicussion \href{https://github.com/cimentadaj/ess/issues/11}{here}.
-For this particular argument, 'sas' is not supported because the data formats have
+For this particular argument in, 'sas' is not supported because the data formats have
 changed between ESS waves and separate formats require different functions to be
 read. To preserve parsimony and format errors between waves, the user should use
 'spss' or 'stata'.
