### 2018-03-25 #201 @maelle
Thanks @nujcharee and thanks @cimentadaj!

I still need to perform the last editor checks (should lead to minor comments only). Maybe it's best to have waited for you to have finished all review changes before I have a last look anyway, so ping me when it's done if that's ok with you?

We could transfer even before the name change, I _think_ redirections from the old account/reponame would still work if it is newaccount/newreponame. But I haven't checked that yet.
