### 2018-03-25 ropensci/essurvey#22 @cimentadaj
Thanks @leeper ! I implemented the change in de3530413a32a6d86950bcdd23c22e7761d258f3 and 90ee16645cbd552a2b1c07f501590b3610422601. The package is now renamed and I'll submit to CRAN soon.
