### 2018-03-25 9e132b8
Remove old tests for ess_* functions
