### 2018-03-25 e37fe1d
Add notes about ess_ deprecated on readme/vignette
