### 2018-03-25 #201 @cimentadaj
Ok @maelle , the changes are in and the package has be renamed. I switched it to `[essurvey](https://github.com/cimentadaj/essurvey)` on Github and everything on the package has been renamed to `essurvey` over `ess`. All tests [pass](https://travis-ci.org/cimentadaj/essurvey/builds/358061542) so I'm just waiting for the badge update. I will submit to CRAN once this review process is over to make the process much smoother.
