### 2018-03-25 #201 @maelle
Approved! Great work @cimentadaj and great reviewer work @leeper & @nujcharee ! A few last comments/questions from me

* Have you checked whether it's ok to use the ESS logo?

* Only a _suggestion_, you could choose to add the reviewers to DESCRIPTION as "rev", with their permission. See [this recent blog post](https://ropensci.org/blog/2018/03/16/thanking-reviewers-in-metadata/)

* the `install.packages` instruction is currently wrong, maybe you should remove it and add it back when the package is on CRAN?

* Add the pkgdown link to the repo URL field and also to DESCRIPTION (after a comma)

Now here is the list of things you have to do before I close this issue 😉
- [] Transfer the repo to the rOpenSci organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  I'll make you an admin once you do.
- [] Fix any links in badges for CI and coverage to point to the ropensci URL
- [] add rOpenSci footer to README
`[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)`
- [] In your .github dir can you also add a CONTRIBUTING.md and a PR template, see https://github.com/ropensci/dotgithubfiles/ for examples
- [] We're starting to roll out software metadata files to all ropensci packages via the Codemeta initiative, see https://github.com/ropensci/codemetar/#codemetar for how to include it in your package, after installing the package - should be easy as running codemetar::write_codemeta() in the root of your package

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
