### 2018-03-26 #201 @cimentadaj
@maelle thank you very much for the review. Thanks as well to@leeper and @nujcharee for some great reviews. `essurvey` benefited a lot from your comments.

Answer's to @maelle's comments below:

Have you checked whether it's ok to use the ESS logo?
Hmm, I'm not entirely sure but the communications office of the ESS has been retweeting the package constantly in their official account, so I'm guessing there shouldn't be a problem. Here's [one example](https://twitter.com/ESS_Survey/status/951866435130798081) but there's plenty more.

Only a suggestion, you could choose to add the reviewers to DESCRIPTION as "rev", with their permission. See this recent blog post

Of course! Already implemented. @nujcharee is your complete name Nujcharee Haswell? Let me know in case I need to fix it from the DESCRIPTION file.

the install.packages instruction is currently wrong, maybe you should remove it and add it back when the package is on CRAN? Yeah, I forgot, thanks!

Add the pkgdown link to the repo URL field and also to DESCRIPTION (after a comma)
I added the pkg website to DESCRIPTION for I can't seem to find a URL field in the repository. What do you mean exactly?
