### 2018-03-26 848aedb
Final ropensci changes from maelle
