+++ DESCRIPTION
@@ -1,12 +1,16 @@
 Package: essurvey
 Title: Download Data from the European Social Survey on the Fly
 Version: 1.0.0
-Authors@R: person("Jorge", "Cimentada", email = "cimentadaj@gmail.com", role = c("aut", "cre"))
+Authors@R: c(
+    person("Jorge", "Cimentada", email = "cimentadaj@gmail.com", role = c("aut", "cre")),   
+    person("Thomas", "Leeper", role = "rev", comment = "Thomas reviewed the package for rOpensci,see https://github.com/ropensci/onboarding/issues/201"),
+    person("Nujcharee", "Haswell", role = "rev", comment = "Nujcharee reviewed the package for rOpensci, see https://github.com/ropensci/onboarding/issues/201")
+  )
 BugReports: https://github.com/cimentadaj/essurvey/issues
 Description: Download data from the European Social Survey directly from their website <http://www.europeansocialsurvey.org/>. There are two family of functions that allow you to download and interactively check all countries and rounds available.
 Depends: R (>= 3.4.0)
 License: MIT + file LICENSE
-URL: https://github.com/cimentadaj/essurvey
+URL: https://github.com/cimentadaj/essurvey, https://cimentadaj.github.io/essurvey/
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
