### 2018-03-26 5ac096f
Final changes to move to ropensci
