+++ NEWS.md
@@ -4,53 +4,53 @@ The `ess` package has been renamed to `essurvey` for a name conflict with Emacs 
 
 ### Breaking changes
 
-* `ess_rounds` and `ess_all_rounds` are deprecated and will be removed in the next release. Use `import_rounds` instead [#22](https://github.com/cimentadaj/ess/issues/22)
+* `ess_rounds` and `ess_all_rounds` are deprecated and will be removed in the next release. Use `import_rounds` instead [#22](https://github.com/ropensci/ess/issues/22)
 
-* `ess_country` and `ess_all_cntrounds` are deprecated and will be removed in the next release. Use `import_countries` instead [#22](https://github.com/cimentadaj/ess/issues/22)
+* `ess_country` and `ess_all_cntrounds` are deprecated and will be removed in the next release. Use `import_countries` instead [#22](https://github.com/ropensci/ess/issues/22)
 
-* The `your_email` argument name of `ess_*` functions has be changed to `ess_email` [#23](https://github.com/cimentadaj/ess/issues/23)
+* The `your_email` argument name of `ess_*` functions has be changed to `ess_email` [#23](https://github.com/ropensci/ess/issues/23)
 
 ### New features
 
 * `import_rounds`, `import_all_rounds` and `download_rounds` have been introduced as
 replacements of `ess_rounds` and `ess_all_rounds`. Same changes were repeated for
-`ess_country` and `ess_all_cntrounds` [#22](https://github.com/cimentadaj/ess/issues/22)
+`ess_country` and `ess_all_cntrounds` [#22](https://github.com/ropensci/ess/issues/22)
 
-* `set_email` to set your email as environmental variable rather than write it in each call [#23](https://github.com/cimentadaj/ess/issues/23)
+* `set_email` to set your email as environmental variable rather than write it in each call [#23](https://github.com/ropensci/ess/issues/23)
 
-* All requests to the ESS website are now done through HTTPS rather than HTTP [#24](https://github.com/cimentadaj/ess/issues/24)
+* All requests to the ESS website are now done through HTTPS rather than HTTP [#24](https://github.com/ropensci/ess/issues/24)
 
-* Add package level documentation [#20](https://github.com/cimentadaj/ess/issues/20)
+* Add package level documentation [#20](https://github.com/ropensci/ess/issues/20)
 
 ### Minor changes
 
-* `ess_email` had no default value but now has `NULL` as default [#23](https://github.com/cimentadaj/ess/issues/23)
+* `ess_email` had no default value but now has `NULL` as default [#23](https://github.com/ropensci/ess/issues/23)
 
-* The `format` argument is now checked through `match.arg` rathern than manual check [#25](https://github.com/cimentadaj/ess/issues/25)
+* The `format` argument is now checked through `match.arg` rathern than manual check [#25](https://github.com/ropensci/ess/issues/25)
 
 ## ess 0.1.1 (2018-03-05)
 
 ### Breaking changes
 
-* Downloading 1 round both for countries or single rounds now returns a data frame rather than a list. If download is more than two rounds it returns a list. [#8](https://github.com/cimentadaj/ess/issues/8)
+* Downloading 1 round both for countries or single rounds now returns a data frame rather than a list. If download is more than two rounds it returns a list. [#8](https://github.com/ropensci/ess/issues/8)
 
 ### New features
 
-* remove_missings() together with remove_numeric_missings() and remove_character_missings() now allow you to recode the typical categories 'Not applicable', 'Don't know', etc.. into NA's. See the vignette example for more details. [#1](https://github.com/cimentadaj/ess/issues/1)
[…]

[…]
