### 2018-03-26 #201 @cimentadaj
@maelle, finalized all changes to close the issue. You can check out the repo at https://github.com/ropensci/essurvey

Thank you all for this experience! I learned a lot from it and I'd love to write an extended blog post on the review process and how to use the package @stefaniebutland.
