### 2018-03-26 b5e8ca7
Add cran comments to changing name
