### 2018-03-26 7370af5
Update issue_template.md
