+++ man/ess_all_rounds.Rd
@@ -10,7 +10,7 @@ ess_all_rounds(ess_email = NULL, only_download = FALSE,
 \arguments{
 \item{ess_email}{a character vector with your email, such as "your_email@email.com".
 If you haven't registered in the ESS website, create an account at 
-\url{http://www.europeansocialsurvey.org/user/new}. A prefered method is to login
+\url{http://www.europeansocialsurvey.org/user/new}. A preferred method is to login
 through \code{\link{set_email}}.}
 
 \item{only_download}{whether to only download the files as Stata files. Defaults to FALSE.}
