### 2018-03-26 #201 @maelle
@cimentadaj you couldn't see the repo description "edit" button because you were not an admin yet. Now I've made you admin, and I've added the pkgdown URL in the description at the same time, hope that's ok.
