### 2018-03-27 #201 @cimentadaj
@leeper would you also like to be listed as a reviewer in the CRAN submission?
