### 2018-03-27 #201 @maelle
Closing this now that the package has been onboarded. @leeper can still  answer about being listed as "rev" in DESCRIPTION ☺
