### 2018-04-03 #201 @cimentadaj
Hi @leeper, just a reminder, would you like to be listed a reviewer when I submit the package to CRAN?
